backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg[binary]>=3.2.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
email-validator>=2.2.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
python-multipart>=0.0.12
slowapi>=0.1.9
pytest>=8.3.0
httpx>=0.27.0
